feat(language-switcher): moved locale swithcer into one file
